fix(tests): fix import in tests
